Specifying privacy to my bio signals and personal life
Asking clearly for people to not infringe on my private life and bio signals including mind reading and inner voice. I am aware of such methods as being used to manipulate and hurt others and I do not support this ethically morally and take personal issue in this happening without my concern and against my will.